download-notes-file

+++ src/main/java/org/anand/mynoteapp/repository/FileRepository.java
@@ -3,5 +3,5 @@ package org.anand.mynoteapp.repository;
 import org.anand.mynoteapp.entity.FileDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<FileDetails, Long> {
+public interface FileRepository extends JpaRepository<FileDetails, Integer> {
 }
